Remove projectsDependencyGraph from internal spec (#125)
* Remove projectsDependencyGraph from internal spec

The GatewayApi.projectsDependencyGraph was originally used in the Domino
CLI but was removed due to incompatibility with prior code. This command
was not used in any of our scripts or tested until ported over for the
Java client testing. Since it is not used downstream anywhere I am opting
to remove it from the spec to reduce potential burden later.

* Remove unused imports